Start refactoring. fastfield collection is still not generic.
