+++ quickwit-serve/src/rest.rs
@@ -229,9 +229,8 @@ async fn export_endpoint<TSearchService: SearchService>(
         output_format: search_request.output_format.to_string(),
     };
 
-    let response_receiver = search_service.root_export(export_request).await?;
-    let stream = ReceiverStream::new(response_receiver);
-    let body = hyper::Body::wrap_stream(stream);
+    let data = search_service.root_export(export_request).await?;
+    let body = hyper::Body::from(data);
     Ok(body)
 }
 
@@ -531,20 +530,11 @@ mod tests {
     #[tokio::test]
     async fn test_rest_export_api() -> anyhow::Result<()> {
         let mut mock_search_service = MockSearchService::new();
-        let (result_sender, result_receiver) = tokio::sync::mpsc::channel(2);
-        result_sender
-            .send(Ok(bytes::Bytes::from("first row\n")))
-            .await?;
-        result_sender
-            .send(Ok(bytes::Bytes::from("second row")))
-            .await?;
         mock_search_service
             .expect_root_export()
-            .return_once(|_| Ok(result_receiver));
+            .return_once(|_| Ok(bytes::Bytes::from("first row\nsecond row")));
         let rest_export_api_handler =
             super::export_handler(Arc::new(mock_search_service)).recover(recover_fn);
-        // send needs to be dropped so otherwise we will wait indefinititely new rows.
-        drop(result_sender);
         let response = warp::test::request()
             .path("/api/v1/my-index/export?query=obama&fastField=external_id&outputFormat=csv")
             .reply(&rest_export_api_handler)
